[TM-1578] Prevent adding a species that is already on the list.
