+++ src/components/elements/Inputs/TreeSpeciesInput/TreeSpeciesInput.tsx
@@ -6,6 +6,8 @@ import { FieldError, FieldErrors } from "react-hook-form";
 import { Else, If, Then, When } from "react-if";
 import { v4 as uuidv4 } from "uuid";
 
+import NonScientificConfirmationModal from "@/components/elements/Inputs/TreeSpeciesInput/NonScientificConfirmationModal";
+import SpeciesAlreadyExistsModal from "@/components/elements/Inputs/TreeSpeciesInput/SpeciesAlreadyExistsModal";
 import { useAutocompleteSearch } from "@/components/elements/Inputs/TreeSpeciesInput/useAutocompleteSearch";
 import Icon, { IconNames } from "@/components/extensive/Icon/Icon";
 import List from "@/components/extensive/List/List";
@@ -42,45 +44,6 @@ export interface TreeSpeciesInputProps extends Omit<InputWrapperProps, "error"> 
 }
 
 export type TreeSpeciesValue = { uuid?: string; name?: string; taxon_id?: string; amount?: number };
-
-const NonScientificConfirmationModal = ({ onConfirm }: { onConfirm: () => void }) => {
-  const t = useT();
-  const { closeModal } = useModalContext();
-
-  return (
-    <div className="margin-4 z-50 m-auto flex max-h-full flex-col items-center justify-start overflow-y-auto rounded-lg border-2 border-neutral-100 bg-white">
-      <div className="flex w-full items-center justify-center gap-1 border-b-2 border-neutral-100 py-1">
-        <Icon name={IconNames.EXCLAMATION_CIRCLE_FILL} className="min-h-4 min-w-4 mb-1 h-4 w-4 text-tertiary-600" />
-        <Text variant="text-16-semibold" className="mb-1 text-blueCustom-700">
-          {t("Your input is a not a scientific name")}
-        </Text>
-      </div>
-      <div className="w-full p-4">
-        <div className="w-full rounded-lg border border-dashed bg-neutral-250 p-2">
-          <div className="flex items-center gap-1">
-            <Text variant="text-14-light" className="text-blueCustom-700">
-              {t("You can add this species, but it will be pending review from Admin.")}
-            </Text>
-          </div>
-        </div>
-        <div className="mt-4 flex w-full justify-end gap-3">
-          <Button variant="secondary" onClick={() => closeModal(ModalId.ERROR_MODAL)}>
-            {t("CANCEL")}
-          </Button>
-          <Button
-            variant="primary"
-            onClick={() => {
-              closeModal(ModalId.ERROR_MODAL);
-              onConfirm();
-            }}
-          >
-            {t("CONFIRM")}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const TreeSpeciesInput = (props: TreeSpeciesInputProps) => {
   const id = useId();
@@ -183,8 +146,11 @@ const TreeSpeciesInput = (props: TreeSpeciesInputProps) => {
       lastInputRef.current?.focus();
     };
 
-    if (!isEmpty(searchResult) && taxonId == null) {
-      // In this case the use had valid values to choose from, but decided to add a value that isn't
+    if (value.find(({ name }) => name === valueAutoComplete) != null) {
+      openModal(ModalId.ERROR_MODAL, <SpeciesAlreadyExistsModal speciesName={valueAutoComplete} />);
+      setValueAutoComplete("");
+    } else if (!isEmpty(searchResult) && taxonId == null) {
+      // In this case the user had valid values to choose from, but decided to add a value that isn't
       // on the list, so they haven't been shown the warning yet.
       openModal(ModalId.ERROR_MODAL, <NonScientificConfirmationModal onConfirm={doAdd} />);
     } else {
